+++ doc/about/changes.md
@@ -7,6 +7,10 @@ in progress
 ### New features
 - Add support for system headers when compiling with MSVC
 
+### Bug fixes
+- Automatically update dependencies via `mopack` if the toolchain file has
+  changed
+
 ---
 
 ## v0.7.1